improve git state resolution
