+++ gantry/util.py
@@ -5,11 +5,9 @@ import time
 from dataclasses import asdict, dataclass
 from datetime import timedelta
 from enum import Enum
-from functools import cached_property
 from pathlib import Path
-from typing import Optional, Tuple, cast
+from typing import Optional, cast
 
-import requests
 import rich
 from beaker import (
     Beaker,
@@ -27,8 +25,6 @@ from beaker.exceptions import (
     BeakerSecretNotFound,
     BeakerWorkspaceNotSet,
 )
-from git.cmd import Git
-from git.repo import Repo
 from rich import print, prompt
 from rich.console import Console
 from rich.status import Status
@@ -263,91 +259,6 @@ def display_results(beaker: Beaker, workload: BeakerWorkload, job: BeakerJob):
         raise ValueError(f"unexpected workload status '{status}'")
 
 
-def parse_git_remote_url(url: str) -> Tuple[str, str]:
-    """
-    Parse a git remote URL into a GitHub (account, repo) pair.
-
-    :raises InvalidRemoteError: If the URL can't be parsed correctly.
-    """
-    try:
-        account, repo = (
-            url.split("https://github.com/")[-1]
-            .split("git@github.com:")[-1]
-            .split(".git")[0]
-            .split("/")
-        )
-    except ValueError:
-        raise InvalidRemoteError(f"Failed to parse GitHub repo path from remote '{url}'")
-    return account, repo
-
-
-@dataclass
-class GitConfig:
-    repo: str
-    repo_url: str
-    ref: str
-    branch: Optional[str] = None
-
-    @property
-    def is_dirty(self) -> bool:
-        repo = Repo(".")
-        return repo.is_dirty()
-
-    @cached_property
-    def is_public(self) -> bool:
-        response = requests.get(self.repo_url)
-        if response.status_code not in {200, 404}:
-            response.raise_for_status()
-        return response.status_code == 200
-
-    @cached_property
-    def ref_exists_on_remote(self) -> bool:
-        git = Git(".")
-        output = cast(
-            str, git.execute(["git", "branch", "-r", "--contains", self.ref], stdout_as_string=True)
-        )
-        output = output.strip()
-        return len(output) > 0
-
-    @property
-    def ref_url(self) -> str:
-        return f"{self.repo_url}/commit/{self.ref}"
-
-    @property
-    def branch_url(self) -> Optional[str]:
-        if self.branch is None:
-            return None
-        else:
-            return f"{self.repo_url}/tree/{self.branch}"
-
-    @classmethod
-    def from_env(cls, ref: Optional[str] = None) -> "GitConfig":
-        repo = Repo(".")
-
-        git_ref = ref or str(repo.commit())
-        remote = repo.remote()
-
-        try:
-            branch = repo.active_branch.tracking_branch()
-        except TypeError:
-            branch = None
-
-        branch_name: Optional[str] = None
-        if branch is not None:
-            remote = repo.remote(branch.remote_name)
-            assert branch.name.startswith(branch.remote_name + "/")
-            branch_name = branch.name.replace(branch.remote_name + "/", "", 1)
-
-        account, repo_name = parse_git_remote_url(remote.url)
-
-        return cls(
-            repo=f"{account}/{repo_name}",
-            repo_url=f"https://github.com/{account}/{repo_name}",
-            ref=git_ref,
-            branch=branch_name,
-        )
-
-
 def ensure_entrypoint_dataset(beaker: Beaker) -> BeakerDataset:
     import hashlib
     from importlib.resources import read_binary
@@ -509,7 +420,7 @@ def init_client(
     yes: bool = False,
     ensure_workspace: bool = True,
 ) -> Beaker:
-    Beaker.MAX_RETRIES = 100  # basically retry forever.
+    Beaker.MAX_RETRIES = 100  # type: ignore
 
     beaker = (
         Beaker.from_env() if workspace is None else Beaker.from_env(default_workspace=workspace)
